Install the event handler in the QCefView constructor
While we install our event handler onto our window when the parent changes later on, we previously didn't install it when we initially construct the QCefView.

If we never changed the QCefView's parent, the notifyMoveOrResizeStarted() method would never be called, causing some strange behaviour (e.g. dropdowns not appearing in the correct place after the window is moved).

This has been addressed by installing the event filter on construction.

Reform the build system

Add support for appveyor